builtin_plugins/vera_google_sheets_report/pyproject.toml
[project]
name = "vera-google-sheets-report"
version = "1.0.2"
description = "Send vera's report to Google Sheets"
readme = "README.md"
authors = [
    { name = "Tal Shapir", email = "talshapir@google.com" },
]
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "vera>=1.0.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "ty>=0.0.4",
]

[project.entry-points.vera]
vera_google_sheets_report = "vera_google_sheets_report.plugin_impl"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
vera = { path = "../../vera" }

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
